Creation of Search Service. So all the search questions will be handeld by a seprarted Service
